+++ src/Components/OpenStackObjectStorageSystem.php
@@ -6,17 +6,20 @@ use DreamFactory\Core\Enums\HttpStatusCodes;
 use DreamFactory\Core\Exceptions\DfException;
 use DreamFactory\Core\Exceptions\BadRequestException;
 use DreamFactory\Core\Exceptions\NotFoundException;
+use DreamFactory\Core\Exceptions\RestException;
 use DreamFactory\Core\File\Components\RemoteFileSystem;
 use DreamFactory\Core\Utility\Session;
 use DreamFactory\Core\Utility\FileUtilities;
+use Guzzle\Http\Exception\BadResponseException;
+use Guzzle\Http\Exception\ClientErrorResponseException;
 use InvalidArgumentException;
 use OpenCloud\Common\Exceptions\ObjFetchError;
-use OpenCloud\Common\Request\Response\Http;
+use OpenCloud\Rackspace;
+use OpenCloud\OpenStack;
 use OpenCloud\Common\Collection;
 use OpenCloud\ObjectStore\Service;
-use OpenCloud\ObjectStore\Container;
-use OpenCloud\ObjectStore\DataObject;
-use OpenCloud\Common\Exceptions\ContainerNotFoundError;
+use OpenCloud\ObjectStore\Resource\Container;
+use OpenCloud\ObjectStore\Resource\DataObject;
 
 /**
  * Class OpenStackObjectStorageSystem
@@ -93,21 +96,22 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
 
         try {
             if (empty($apiKey)) {
-                $os = new DfOpenStack($authUrl, $secret);
+                $os = new OpenStack($authUrl, $secret);
             } else {
                 $pos = stripos($authUrl, '/v');
                 if (false !== $pos) {
                     $authUrl = substr($authUrl, 0, $pos);
                 }
                 $authUrl = FileUtilities::fixFolderPath($authUrl) . 'v2.0';
-                $os = new DfRackspace($authUrl, $secret);
+                $os = new Rackspace($authUrl, $secret);
             }
 
-            $this->blobConn = $os->ObjectStore('cloudFiles', $region);
+            $this->blobConn = $os->objectStoreService('cloudFiles', $region);
             if (!$this->containerExists($this->container)) {
                 $this->createContainer(['name' => $this->container]);
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException('Failed to launch OpenStack service: ' . $ex->getMessage());
         }
     }
@@ -150,6 +154,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
 
             return $out;
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException('Failed to list containers: ' . $ex->getMessage());
         }
     }
@@ -185,11 +190,10 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
 
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             $result['size'] = $container->bytes;
-        } catch (ContainerNotFoundError $ex) {
-            throw new DfException('Failed to find container: ' . $ex->getMessage());
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException('Failed to get container: ' . $ex->getMessage());
         }
 
@@ -210,12 +214,16 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
 
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
 
             return !empty($container);
-        } catch (ContainerNotFoundError $ex) {
-            return false;
         } catch (\Exception $ex) {
+            if ($ex instanceof ClientErrorResponseException) {
+                if ($ex->getCode() === 404 || str_contains($ex->getMessage(), '404')) {
+                    return false;
+                }
+            }
+            static::handleGuzzleException($ex);
             throw new DfException('Failed to list containers: ' . $ex->getMessage());
         }
     }
@@ -238,15 +246,13 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
             throw new BadRequestException('No name found for container in create request.');
         }
         try {
-            /** @var Container $container */
-            $container = $this->blobConn->Container();
-            $params = ['name' => $name];
-            if (!$container->Create($params)) {
+            if (!$container = $this->blobConn->createContainer($name)) {
                 throw new \Exception('');
             }
 
             return ['name' => $name, 'path' => $name];
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to create container '$name': " . $ex->getMessage());
         }
     }
@@ -265,7 +271,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
@@ -274,6 +280,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
                 throw new \Exception('');
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to update container '$container': " . $ex->getMessage());
         }
     }
@@ -293,7 +300,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
@@ -302,6 +309,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
                 throw new \Exception('');
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to delete container '$container': " . $ex->getMessage());
         }
     }
@@ -320,7 +328,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
@@ -343,26 +351,28 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
      * @throws DfException
      * @throws \Exception
      */
-    public function putBlobData($container = '', $name = '', $blob = '', $type = '')
+    public function putBlobData($container = '', $name = '', $blob = null, $type = '')
     {
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
 
+            /** @var DataObject $obj */
             $obj = $container->DataObject();
-            $obj->SetData($blob);
-            $obj->name = $name;
+            $obj->setContent($blob);
+            $obj->setName($name);
             if (!empty($type)) {
-                $obj->content_type = $type;
+                $obj->setContentType($type);
             }
-            if (!$obj->Create()) {
+            if (!$obj->update()) {
                 throw new \Exception('');
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to create blob '$name': " . $ex->getMessage());
         }
     }
@@ -376,26 +386,24 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
      * @throws DfException
      * @throws \Exception
      */
-    public function putBlobFromFile($container = '', $name = '', $localFileName = '', $type = '')
+    public function putBlobFromFile($container = '', $name = '', $localFileName = null, $type = '')
     {
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
 
-            $obj = $container->DataObject();
             $params = ['name' => $name];
             if (!empty($type)) {
                 $params['content_type'] = $type;
             }
 
-            if (!$obj->Create($params, $localFileName)) {
-                throw new \Exception('');
-            }
+            $container->uploadObject($name, file_get_contents($localFileName), $params);
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to create blob '$name': " . $ex->getMessage());
         }
     }
@@ -415,22 +423,23 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $src_container */
-            $src_container = $this->blobConn->Container($src_container);
+            $src_container = $this->blobConn->getContainer($src_container);
             if (empty($src_container)) {
                 throw new \Exception("No container named '$src_container'");
             }
             /** @var Container $dest_container */
-            $dest_container = $this->blobConn->Container($container);
+            $dest_container = $this->blobConn->getContainer($container);
             if (empty($dest_container)) {
                 throw new \Exception("No container named '$container'");
             }
 
             $source = $src_container->DataObject($src_name);
             $destination = $dest_container->DataObject();
-            $destination->name = $name;
+            $destination->setName($name);
 
             $source->Copy($destination);
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to copy blob '$name': " . $ex->getMessage());
         }
     }
@@ -445,12 +454,12 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
      * @throws DfException
      * @throws \Exception
      */
-    public function getBlobAsFile($container = '', $name = '', $localFileName = '')
+    public function getBlobAsFile($container = '', $name = '', $localFileName = null)
     {
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
@@ -461,6 +470,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
                 throw new \Exception('');
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to retrieve blob '$name': " . $ex->getMessage());
         }
     }
@@ -478,15 +488,19 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
 
-            $obj = $container->DataObject($name);
+            $obj = $container->getObject($name);
+            $request = $obj->getClient()->get($obj->getUrl());
+            /** @var \Guzzle\Http\Message\Response $result */
+            $result = $request->send();
 
-            return $obj->SaveToString();
+            return $result->getBody(true);
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException("Failed to retrieve blob '$name': " . $ex->getMessage());
         }
     }
@@ -504,7 +518,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
@@ -522,6 +536,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
                 $obj->Delete();
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             if ($ex instanceof ObjFetchError) {
                 throw new NotFoundException("File '$name' was not found.'");
             }
@@ -553,7 +568,7 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         }
 
         /** @var Container $container */
-        $container = $this->blobConn->Container($container);
+        $container = $this->blobConn->getContainer($container);
         if (empty($container)) {
             throw new \Exception("No container named '$container'");
         }
@@ -564,21 +579,21 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $out = [];
 
         /** @var DataObject $obj */
-        while (($obj = $list->Next())) {
-            if (!empty($obj->name)) {
-                if (0 == strcmp($prefix, $obj->name)) {
+        while ($obj = $list->Next()) {
+            if (!empty($obj->getName()) && $obj->isDirectory() === false) {
+                if (0 == strcmp($prefix, $obj->getName())) {
                     continue;
                 }
                 $out[] = [
-                    'name'           => $obj->name,
-                    'content_type'   => $obj->content_type,
-                    'content_length' => $obj->bytes,
-                    'last_modified'  => gmdate('D, d M Y H:i:s \G\M\T', strtotime($obj->last_modified))
+                    'name'           => $obj->getName(),
+                    'content_type'   => $obj->getContentType(),
+                    'content_length' => $obj->getContentLength(),
+                    'last_modified'  => gmdate('D, d M Y H:i:s \G\M\T', strtotime($obj->getLastModified()))
                 ];
-            } elseif (!empty($obj->subdir)) // sub directories formatted differently
+            } elseif (!empty($obj->getDirectory())) // sub directories formatted differently
             {
                 $out[] = [
-                    'name'           => $obj->subdir,
+                    'name'           => $obj->getName(),
                     'content_type'   => null,
                     'content_length' => 0,
                     'last_modified'  => null
@@ -604,22 +619,30 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
 
             $obj = $container->DataObject($name);
 
-            $file = [
-                'name'           => $obj->name,
-                'content_type'   => $obj->content_type,
-                'content_length' => $obj->bytes,
-                'last_modified'  => gmdate('D, d M Y H:i:s \G\M\T', strtotime($obj->last_modified))
-            ];
+            if ($obj->getName() === null && $obj->isDirectory() === false) {
+                // Container itself here
+                $file = [
+                    'name' => '.'
+                ];
+            } else {
+                $file = [
+                    'name'           => $obj->getName(),
+                    'content_type'   => $obj->getContentType(),
+                    'content_length' => $obj->getContentLength(),
+                    'last_modified'  => gmdate('D, d M Y H:i:s \G\M\T', strtotime($obj->getLastModified()))
+                ];
+            }
 
             return $file;
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             throw new DfException('Failed to list metadata: ' . $ex->getMessage());
         }
     }
@@ -637,16 +660,16 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
         $this->checkConnection();
         try {
             /** @var Container $container */
-            $container = $this->blobConn->Container($container);
+            $container = $this->blobConn->getContainer($container);
             if (empty($container)) {
                 throw new \Exception("No container named '$container'");
             }
 
             $obj = $container->DataObject($name);
 
-            header('Last-Modified: ' . $obj->last_modified);
-            header('Content-Type: ' . $obj->content_type);
-            header('Content-Length:' . $obj->content_length);
+            header('Last-Modified: ' . $obj->getLastModified());
+            header('Content-Type: ' . $obj->getContentType());
+            header('Content-Length:' . $obj->getContentLength());
 
             $disposition =
                 (isset($params['disposition']) && !empty($params['disposition'])) ? $params['disposition']
@@ -654,21 +677,23 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
 
             header('Content-Disposition: ' . $disposition . '; filename="' . $name . '";');
             $index = 0;
-            $size = (integer)$obj->content_length;
+            $size = (integer)$obj->getContentLength();
             $chunk = \Config::get('df.file_chunk_size');
             ob_clean();
 
             while ($index < $size) {
                 $header = ['Range' => 'bytes=' . $index . '-' . ($index + $chunk - 1)];
-                /** @var Http $result */
-                $result = $container->Service()->Request($obj->Url(), 'GET', $header);
-                $info = $result->info();
-                $length = array_get($info, 'size_download');
+                /** @var \Guzzle\Http\Message\RequestInterface $result */
+                $request = $container->getService()->getClient()->get($obj->getUrl(), $header);
+                /** @var \Guzzle\Http\Message\Response $result */
+                $result = $request->send();
+                $length = $result->getContentLength();
                 $index += $length;
                 flush();
-                echo $result->HttpBody();
+                echo $result->getBody();
             }
         } catch (\Exception $ex) {
+            static::handleGuzzleException($ex);
             if ($ex instanceof ObjFetchError) {
                 $code = HttpStatusCodes::HTTP_NOT_FOUND;
                 $status_header = "HTTP/1.1 $code";
@@ -679,5 +704,67 @@ class OpenStackObjectStorageSystem extends RemoteFileSystem
                 throw new DfException('Failed to stream blob: ' . $ex->getMessage());
             }
         }
+    }
+
+    /**
+     * @param $ex
+     *
+     * @throws \DreamFactory\Core\Exceptions\BadRequestException
+     * @throws \DreamFactory\Core\Exceptions\NotFoundException
+     * @throws \DreamFactory\Core\Exceptions\RestException
+     */
+    public static function handleGuzzleException($ex)
+    {
+        if ($ex instanceof BadResponseException) {
+            $code = static::getGuzzleExceptionCode($ex);
+            $message = static::cleanGuzzleExceptionMessage($ex);
+            switch ($code) {
+                case 404:
+                    throw new NotFoundException($message);
+                case 400:
+                    throw new BadRequestException($message);
+                default:
+                    throw new RestException($code, $message);
+            }
+        }
+    }
+
+    /**
+     * @param \Guzzle\Http\Exception\BadResponseException $ex
+     *
+     * @return int
+     */
+    public static function getGuzzleExceptionCode(BadResponseException $ex)
+    {
+        $msgs = explode("\n", $ex->getMessage());
+        foreach ($msgs as $msg) {
+            if ($msg !== $code = str_replace('[status code]', '', $msg)) {
+                return (integer)trim($code);
+            }
+        }
+    }
+
+    /**
+     * @param \Guzzle\Http\Exception\BadResponseException $ex
+     *
+     * @return string
+     */
+    public static function cleanGuzzleExceptionMessage(BadResponseException $ex)
+    {
+        $message = '';
+        $msgs = explode("\n", $ex->getMessage());
+        foreach ($msgs as $msg) {
+            if ($msg !== $code = str_replace('[status code]', '', $msg)) {
+                $message = '[' . trim($code) . '] ' . $message;
+            } elseif ($msg !== $phrase = str_replace('[reason phrase]', '', $msg)) {
+                $message .= (empty($message)) ? trim($phrase) : '. ' . trim($phrase);
+            } elseif ($msg !== $url = str_replace('[url]', '', $msg)) {
+                $message .= (empty($message)) ? trim($url) : '. ' . trim($url);
+            } else {
+                $message .= (empty($message)) ? $msg : '. ' . $msg;
+            }
+        }
+
+        return $message;
     }
 }